+++ merlict-c89-view.c
@@ -8,7 +8,9 @@
 #include <stdio.h>
 
 #include "merlict_c89.h"
+#include "merlict_c89.c"
 #include "merlict_c89_viewer.h"
+#include "merlict_c89_viewer.c"
 
 int main(int argc, char *argv[])
 {
