+++ merlict_c89_viewer.c
@@ -1,0 +1,4 @@
+#include "merlict_c89_viewer/mlivr_toggle_stdin.c"
+#include "merlict_c89_viewer/mlivr_viewer.c"
+#include "merlict_c89_viewer/mlivrCursor.c"
+#include "merlict_c89_viewer/mlivrConfig.c"
