+++ merlict_c89_viewer/mlivr_version.h
@@ -1,0 +1,9 @@
+/* Copyright 2018-2020 Sebastian Achim Mueller */
+#ifndef MERLICT_C89_VIWER_MLIVR_VERSION_H_
+#define MERLICT_C89_VIWER_MLIVR_VERSION_H_
+
+#define MLIVR_VERSION_MAYOR 0
+#define MLIVR_VERSION_MINOR 0
+#define MLIVR_VERSION_PATCH 0
+
+#endif
